Move non-application-specific utils back to lib/utils.ts
