+++ packages/api/src/client/postsApi.ts
@@ -2,6 +2,7 @@ import { da, render } from '@urbit/aura';
 
 import type { Poke } from '../http-api';
 import { createDevLogger } from '../lib/logger';
+import { IMAGE_URL_REGEX } from '../lib/utils';
 import type * as db from '../types/models';
 import { ContentReference } from '../types/references';
 import * as ub from '../urbit';
@@ -46,7 +47,6 @@ import {
   getTextContent as getTextPreview,
 } from './postContent';
 import { poke, scry, subscribeOnce } from './urbit';
-import { IMAGE_URL_REGEX } from './utils';
 
 const logger = createDevLogger('postsApi', false);
