+++ packages/shared/src/store/lure.ts
@@ -1,7 +1,7 @@
 import { useQuery } from '@tanstack/react-query';
 import { getCurrentUserId, scry, subscribeOnce } from '@tloncorp/api';
 import { DeepLinkMetadata, createDeepLink } from '@tloncorp/api/client/branch';
-import { asyncWithDefault } from '@tloncorp/api/client/utils';
+import { asyncWithDefault } from '@tloncorp/api/lib/utils';
 import produce from 'immer';
 import { useEffect, useMemo, useState } from 'react';
 import create from 'zustand';
